+++ includes/glacial-cpt-acf-main.php
@@ -11,6 +11,13 @@ if ( !defined( 'ABSPATH' ) ) {
 }
 
 /**
+ * Include our ACF helpers and filters
+ *
+ * @since 2.1.1
+ * */
+include GLACIAL_CPT_PLUGIN_DIR . 'includes/glacial-cpt-acf-helpers-filters.php';
+
+/**
  * Check to see if phone modal is enabled
  *
  * @since 2.1.0
@@ -58,6 +65,23 @@ function glacial_cpt_flush_rewrite_rules(): void {
 
 add_action( 'init', 'glacial_cpt_flush_rewrite_rules', 20 );
 
+function glacial_cpt_is_dev_environment(): bool {
+	if ( !defined( 'WP_ENVIRONMENT_TYPE' ) ) {
+		return false;
+	}
+
+	if ( in_array( WP_ENVIRONMENT_TYPE, array( 'local' ), true ) ) {
+		return true;
+	}
+
+	// if .local is in the site url, we are in a dev environment
+	if ( str_contains( get_site_url(), '.local' ) ) {
+		return true;
+	}
+
+	return false;
+}
+
 /**
  * ACF JSON save point
  *
@@ -67,8 +91,10 @@ function glacial_cpt_json_save_point( $acf_json_path ): string {
 	return GLACIAL_CPT_PLUGIN_DIR . '/cpt-acf-json';
 }
 
-// Only save ACF JSON in local and development environments
-if ( WP_ENVIRONMENT_TYPE === 'local' || WP_ENVIRONMENT_TYPE === 'development' ) {
+/*
+ * Only save ACF JSON in dev environments
+ * */
+if ( glacial_cpt_is_dev_environment() ) {
 	add_filter( 'acf/settings/save_json', 'glacial_cpt_json_save_point' );
 }